+++ framework/src/test/java/stest/tron/wallet/dailybuild/operationupdate/MutiSignAccountPermissionUpdateTest.java
@@ -3,9 +3,7 @@ package stest.tron.wallet.dailybuild.operationupdate;
 import com.google.protobuf.ByteString;
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
-import java.util.HashMap;
 import java.util.Optional;
-import java.util.Random;
 import java.util.concurrent.TimeUnit;
 import lombok.extern.slf4j.Slf4j;
 import org.testng.Assert;
@@ -21,24 +19,30 @@ import org.tron.core.Wallet;
 import org.tron.protos.Protocol.TransactionInfo;
 import stest.tron.wallet.common.client.Configuration;
 import stest.tron.wallet.common.client.Parameter.CommonConstant;
-import stest.tron.wallet.common.client.utils.Base58;
 import stest.tron.wallet.common.client.utils.PublicMethed;
 import stest.tron.wallet.common.client.utils.PublicMethedForMutiSign;
 
 @Slf4j
-public class WalletTestMutiSign013 {
+public class MutiSignAccountPermissionUpdateTest {
 
+  private static final long now = System.currentTimeMillis();
+  private static final long totalSupply = now;
+  private static String name = "MutiSign001_" + Long.toString(now);
   private final String testKey002 = Configuration.getByPath("testng.conf")
       .getString("foundationAccount.key1");
   private final byte[] fromAddress = PublicMethed.getFinalAddress(testKey002);
-  private final String witnessKey001 = Configuration.getByPath("testng.conf")
-      .getString("witness.key1");
-  private final byte[] witnessAddress = PublicMethed.getFinalAddress(witnessKey001);
+  private final String testKey001 = Configuration.getByPath("testng.conf")
+      .getString("foundationAccount.key2");
+  private final byte[] fromAddress01 = PublicMethed.getFinalAddress(testKey001);
   private final String operations = Configuration.getByPath("testng.conf")
       .getString("defaultParameter.operations");
+  String description = Configuration.getByPath("testng.conf")
+      .getString("defaultParameter.assetDescription");
+  String url = Configuration.getByPath("testng.conf")
+      .getString("defaultParameter.assetUrl");
   ByteString assetAccountId1;
   String[] permissionKeyString = new String[2];
-  String[] ownerKeyString = new String[3];
+  String[] ownerKeyString = new String[2];
   String accountPermissionJson = "";
   ECKey ecKey1 = new ECKey(Utils.getRandom());
   byte[] manager1Address = ecKey1.getAddress();
@@ -50,28 +54,16 @@ public class WalletTestMutiSign013 {
   byte[] ownerAddress = ecKey3.getAddress();
   String ownerKey = ByteArray.toHexString(ecKey3.getPrivKeyBytes());
   ECKey ecKey4 = new ECKey(Utils.getRandom());
-  byte[] newAddress = ecKey4.getAddress();
-  String newKey = ByteArray.toHexString(ecKey4.getPrivKeyBytes());
+  byte[] participateAddress = ecKey4.getAddress();
+  String participateKey = ByteArray.toHexString(ecKey4.getPrivKeyBytes());
   private long multiSignFee = Configuration.getByPath("testng.conf")
       .getLong("defaultParameter.multiSignFee");
   private long updateAccountPermissionFee = Configuration.getByPath("testng.conf")
       .getLong("defaultParameter.updateAccountPermissionFee");
   private ManagedChannel channelFull = null;
   private WalletGrpc.WalletBlockingStub blockingStubFull = null;
-  private String fullnode = Configuration.getByPath("testng.conf").getStringList("fullnode.ip.list")
-      .get(0);
-
-  /**
-   * constructor.
-   */
-
-  public static byte[] randomBytes(int length) {
-    // generate the random number
-    byte[] result = new byte[length];
-    new Random().nextBytes(result);
-    result[0] = Wallet.getAddressPreFixByte();
-    return result;
-  }
+  private String fullnode = Configuration.getByPath("testng.conf")
+      .getStringList("fullnode.ip.list").get(0);
 
   @BeforeSuite
   public void beforeSuite() {
@@ -83,7 +75,7 @@ public class WalletTestMutiSign013 {
    * constructor.
    */
 
-  @BeforeClass(enabled = false)
+  @BeforeClass(enabled = true)
   public void beforeClass() {
     channelFull = ManagedChannelBuilder.forTarget(fullnode)
         .usePlaintext(true)
@@ -91,8 +83,8 @@ public class WalletTestMutiSign013 {
     blockingStubFull = WalletGrpc.newBlockingStub(channelFull);
   }
 
-  @Test(enabled = false)
-  public void testMutiSignForAccount() {
+  @Test(enabled = true)
+  public void testMutiSign1UpdatePermission() {
     ecKey1 = new ECKey(Utils.getRandom());
     manager1Address = ecKey1.getAddress();
     manager1Key = ByteArray.toHexString(ecKey1.getPrivKeyBytes());
@@ -106,35 +98,24 @@ public class WalletTestMutiSign013 {
     ownerKey = ByteArray.toHexString(ecKey3.getPrivKeyBytes());
     PublicMethed.printAddress(ownerKey);
 
-    ecKey4 = new ECKey(Utils.getRandom());
-    newAddress = ecKey4.getAddress();
-    newKey = ByteArray.toHexString(ecKey4.getPrivKeyBytes());
-
-    long needCoin = updateAccountPermissionFee * 1 + multiSignFee * 10;
+    long needCoin = updateAccountPermissionFee * 2 + multiSignFee * 3;
 
     Assert.assertTrue(
-        PublicMethed.sendcoin(ownerAddress, needCoin + 100000000L, fromAddress, testKey002,
+        PublicMethed.sendcoin(ownerAddress, needCoin, fromAddress, testKey002,
             blockingStubFull));
-
-    Assert.assertTrue(PublicMethed
-        .freezeBalanceForReceiver(fromAddress, 1000000000, 0, 0, ByteString.copyFrom(ownerAddress),
-            testKey002, blockingStubFull));
-
     PublicMethed.waitProduceNextBlock(blockingStubFull);
-
-    Long balanceBefore = PublicMethed.queryAccount(ownerAddress, blockingStubFull)
-        .getBalance();
+    Long balanceBefore = PublicMethed.queryAccount(ownerAddress, blockingStubFull).getBalance();
     logger.info("balanceBefore: " + balanceBefore);
 
     permissionKeyString[0] = manager1Key;
     permissionKeyString[1] = manager2Key;
     ownerKeyString[0] = ownerKey;
     ownerKeyString[1] = manager1Key;
-    ownerKeyString[2] = manager2Key;
+    String[] permissionKeyString1 = new String[1];
+    permissionKeyString1[0] = ownerKey;
     accountPermissionJson =
-        "{\"owner_permission\":{\"type\":0,\"permission_name\":\"owner\",\"threshold\":3,\"keys\":["
+        "{\"owner_permission\":{\"type\":0,\"permission_name\":\"owner\",\"threshold\":2,\"keys\":["
             + "{\"address\":\"" + PublicMethed.getAddressString(manager1Key) + "\",\"weight\":1},"
-            + "{\"address\":\"" + PublicMethed.getAddressString(manager2Key) + "\",\"weight\":1},"
             + "{\"address\":\"" + PublicMethed.getAddressString(ownerKey)
             + "\",\"weight\":1}]},"
             + "\"active_permissions\":[{\"type\":2,\"permission_name\":\"active0\",\"threshold\":2,"
@@ -143,17 +124,16 @@ public class WalletTestMutiSign013 {
             + "{\"address\":\"" + PublicMethed.getAddressString(manager1Key) + "\",\"weight\":1},"
             + "{\"address\":\"" + PublicMethed.getAddressString(manager2Key) + "\",\"weight\":1}"
             + "]}]}";
+
     logger.info(accountPermissionJson);
     String txid = PublicMethedForMutiSign
         .accountPermissionUpdateForTransactionId(accountPermissionJson, ownerAddress, ownerKey,
             blockingStubFull, ownerKeyString);
-
-    final String updateName = Long.toString(System.currentTimeMillis());
     PublicMethed.waitProduceNextBlock(blockingStubFull);
-    Assert.assertNotNull(txid);
 
     Optional<TransactionInfo> infoById = PublicMethed
         .getTransactionInfoById(txid, blockingStubFull);
+
     long balanceAfter = PublicMethed.queryAccount(ownerAddress, blockingStubFull).getBalance();
     long energyFee = infoById.get().getReceipt().getEnergyFee();
     long netFee = infoById.get().getReceipt().getNetFee();
@@ -168,49 +148,44 @@ public class WalletTestMutiSign013 {
     Assert.assertEquals(fee, energyFee + netFee + updateAccountPermissionFee);
 
     balanceBefore = balanceAfter;
-    byte[] accountName = "11z2122328".getBytes();
-    Assert.assertTrue(PublicMethedForMutiSign.createAccount1(
-        ownerAddress, newAddress, ownerKey, blockingStubFull, 0, ownerKeyString));
-    Assert.assertTrue(
-        PublicMethedForMutiSign.setAccountId1(accountName,
-            ownerAddress, ownerKey, 0, blockingStubFull, ownerKeyString));
-    Assert.assertTrue(PublicMethedForMutiSign.sendcoinWithPermissionId(
-        newAddress, 100L, ownerAddress, 0, ownerKey, blockingStubFull, ownerKeyString));
-    Assert.assertTrue(PublicMethedForMutiSign.freezeBalanceWithPermissionId(
-        ownerAddress, 1000000L, 0, 0, ownerKey, blockingStubFull, ownerKeyString));
-    Assert.assertTrue(PublicMethedForMutiSign.freezeBalanceGetEnergyWithPermissionId(
-        ownerAddress, 1000000L, 0, 1, ownerKey, blockingStubFull, 0, ownerKeyString));
-    Assert.assertTrue(PublicMethedForMutiSign.freezeBalanceForReceiverWithPermissionId(
-        ownerAddress, 1000000L, 0, 0, ByteString.copyFrom(newAddress),
-        ownerKey, blockingStubFull, 0, ownerKeyString));
-    Assert.assertTrue(PublicMethedForMutiSign.unFreezeBalanceWithPermissionId(
-        ownerAddress, ownerKey, 0, null, 0, blockingStubFull, ownerKeyString));
-    Assert.assertTrue(PublicMethedForMutiSign.unFreezeBalanceWithPermissionId(
-        ownerAddress, ownerKey, 0, newAddress, 0, blockingStubFull, ownerKeyString));
-    Assert.assertTrue(PublicMethedForMutiSign.updateAccountWithPermissionId(
-        ownerAddress, updateName.getBytes(), ownerKey, blockingStubFull, 0, ownerKeyString));
-
-    String voteStr = Base58.encode58Check(witnessAddress);
-    HashMap<String, String> smallVoteMap = new HashMap<String, String>();
-    smallVoteMap.put(voteStr, "1");
-    Assert.assertTrue(PublicMethedForMutiSign.voteWitnessWithPermissionId(
-        smallVoteMap, ownerAddress, ownerKey, blockingStubFull, 0, ownerKeyString));
-
+    String accountPermissionJson1 =
+        "{\"owner_permission\":{\"type\":0,\"permission_name\":\"owner\",\"threshold\":2,\"keys\":["
+            + "{\"address\":\"" + PublicMethed.getAddressString(manager1Key) + "\",\"weight\":1},"
+            + "{\"address\":\"" + PublicMethed.getAddressString(ownerKey)
+            + "\",\"weight\":1}]},"
+            + "\"active_permissions\":[{\"type\":2,\"permission_name\":\"active0\",\"threshold\":2,"
+            + "\"operations\":\"" + operations + "\","
+            + "\"keys\":["
+            + "{\"address\":\"" + PublicMethed.getAddressString(manager1Key) + "\",\"weight\":1},"
+            + "{\"address\":\"" + PublicMethed.getAddressString(ownerKey) + "\",\"weight\":1}"
+            + "]}]}";
+    txid = PublicMethedForMutiSign
+        .accountPermissionUpdateForTransactionId1(accountPermissionJson1, ownerAddress, ownerKey,
+            blockingStubFull, 0, ownerKeyString);
     PublicMethed.waitProduceNextBlock(blockingStubFull);
 
-    balanceAfter = PublicMethed.queryAccount(ownerAddress, blockingStubFull).getBalance();
-    logger.info("balanceAfter: " + balanceAfter);
-    Assert.assertEquals(balanceBefore - balanceAfter, multiSignFee * 10 + 1000000 + 100);
+    Assert.assertNotNull(txid);
 
-    Assert.assertTrue(
-        PublicMethed.unFreezeBalance(fromAddress, testKey002, 0, ownerAddress, blockingStubFull));
+    infoById = PublicMethed
+        .getTransactionInfoById(txid, blockingStubFull);
+    balanceAfter = PublicMethed.queryAccount(ownerAddress, blockingStubFull).getBalance();
+    energyFee = infoById.get().getReceipt().getEnergyFee();
+    netFee = infoById.get().getReceipt().getNetFee();
+    fee = infoById.get().getFee();
+
+    logger.info("balanceAfter: " + balanceAfter);
+    logger.info("energyFee: " + energyFee);
+    logger.info("netFee: " + netFee);
+    logger.info("fee: " + fee);
+
 
   }
+
 
   /**
    * constructor.
    */
-  @AfterClass(enabled = false)
+  @AfterClass(enabled = true)
   public void shutdown() throws InterruptedException {
     if (channelFull != null) {
       channelFull.shutdown().awaitTermination(5, TimeUnit.SECONDS);
