+++ framework/src/test/java/stest/tron/wallet/dailybuild/operationupdate/MutiSignExchangeContractTest.java
@@ -26,7 +26,7 @@ import stest.tron.wallet.common.client.utils.PublicMethed;
 import stest.tron.wallet.common.client.utils.PublicMethedForMutiSign;
 
 @Slf4j
-public class WalletTestMutiSign016 {
+public class MutiSignExchangeContractTest {
 
   private static final long now = System.currentTimeMillis();
   private static final long totalSupply = 1000000001L;
@@ -87,7 +87,7 @@ public class WalletTestMutiSign016 {
    * constructor.
    */
 
-  @BeforeClass(enabled = false)
+  @BeforeClass(enabled = true)
   public void beforeClass() {
     channelFull = ManagedChannelBuilder.forTarget(fullnode)
         .usePlaintext(true)
@@ -100,7 +100,7 @@ public class WalletTestMutiSign016 {
     blockingStubSolidity = WalletSolidityGrpc.newBlockingStub(channelSolidity);
   }
 
-  @Test(enabled = false, description = "MutiSign for create token")
+  @Test(enabled = true, description = "MutiSign for create token")
   public void test1CreateUsedAsset() {
     ecKey1 = new ECKey(Utils.getRandom());
     exchange001Address = ecKey1.getAddress();
@@ -134,7 +134,7 @@ public class WalletTestMutiSign016 {
     PublicMethed.waitProduceNextBlock(blockingStubFull);
   }
 
-  @Test(enabled = false, description = "MutiSign for create exchange")
+  @Test(enabled = true, description = "MutiSign for create exchange")
   public void test2CreateExchange() {
     ecKey3 = new ECKey(Utils.getRandom());
     manager1Address = ecKey3.getAddress();
@@ -217,7 +217,7 @@ public class WalletTestMutiSign016 {
   }
 
 
-  @Test(enabled = false, description = "Mutisign for inject exchange")
+  @Test(enabled = true, description = "Mutisign for inject exchange")
   public void test4InjectExchange() {
     exchangeIdInfo = PublicMethed.getExchange(exchangeId.toString(), blockingStubFull);
     final Long beforeExchangeToken1Balance = exchangeIdInfo.get().getFirstTokenBalance();
@@ -276,7 +276,7 @@ public class WalletTestMutiSign016 {
     Assert.assertEquals(balanceBefore - balanceAfter, needCoin);
   }
 
-  @Test(enabled = false, description = "MutiSign for withdraw exchange")
+  @Test(enabled = true, description = "MutiSign for withdraw exchange")
   public void test5WithdrawExchange() {
     final long needCoin = multiSignFee;
     Long balanceBefore = PublicMethed.queryAccount(exchange001Address, blockingStubFull)
@@ -338,7 +338,7 @@ public class WalletTestMutiSign016 {
 
   }
 
-  @Test(enabled = false, description = "MutiSign for transaction exchange")
+  @Test(enabled = true, description = "MutiSign for transaction exchange")
   public void test6TransactionExchange() {
     final long needCoin = multiSignFee;
     Long balanceBefore = PublicMethed.queryAccount(exchange001Address, blockingStubFull)
