+++ framework/src/test/java/stest/tron/wallet/dailybuild/operationupdate/MutiSignSmartContractTest.java
@@ -3,7 +3,9 @@ package stest.tron.wallet.dailybuild.operationupdate;
 import com.google.protobuf.ByteString;
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
+import java.util.ArrayList;
 import java.util.Optional;
+import java.util.Random;
 import java.util.concurrent.TimeUnit;
 import lombok.extern.slf4j.Slf4j;
 import org.testng.Assert;
@@ -16,31 +18,30 @@ import org.tron.common.crypto.ECKey;
 import org.tron.common.utils.ByteArray;
 import org.tron.common.utils.Utils;
 import org.tron.core.Wallet;
+import org.tron.protos.Protocol.Block;
 import org.tron.protos.Protocol.TransactionInfo;
+import org.tron.protos.contract.SmartContractOuterClass.SmartContract;
 import stest.tron.wallet.common.client.Configuration;
 import stest.tron.wallet.common.client.Parameter.CommonConstant;
 import stest.tron.wallet.common.client.utils.PublicMethed;
 import stest.tron.wallet.common.client.utils.PublicMethedForMutiSign;
 
 @Slf4j
-public class WalletTestMutiSign011 {
+public class MutiSignSmartContractTest {
 
-  private static final long now = System.currentTimeMillis();
-  private static final long totalSupply = now;
-  private static String name = "MutiSign001_" + Long.toString(now);
   private final String testKey002 = Configuration.getByPath("testng.conf")
       .getString("foundationAccount.key1");
   private final byte[] fromAddress = PublicMethed.getFinalAddress(testKey002);
-  private final String testKey001 = Configuration.getByPath("testng.conf")
-      .getString("foundationAccount.key2");
-  private final byte[] fromAddress01 = PublicMethed.getFinalAddress(testKey001);
   private final String operations = Configuration.getByPath("testng.conf")
       .getString("defaultParameter.operations");
-  String description = Configuration.getByPath("testng.conf")
-      .getString("defaultParameter.assetDescription");
-  String url = Configuration.getByPath("testng.conf")
-      .getString("defaultParameter.assetUrl");
-  ByteString assetAccountId1;
+  ArrayList<String> txidList = new ArrayList<String>();
+  Optional<TransactionInfo> infoById = null;
+  Long beforeTime;
+  Long afterTime;
+  Long beforeBlockNum;
+  Long afterBlockNum;
+  Block currentBlock;
+  Long currentBlockNum;
   String[] permissionKeyString = new String[2];
   String[] ownerKeyString = new String[2];
   String accountPermissionJson = "";
@@ -53,17 +54,18 @@ public class WalletTestMutiSign011 {
   ECKey ecKey3 = new ECKey(Utils.getRandom());
   byte[] ownerAddress = ecKey3.getAddress();
   String ownerKey = ByteArray.toHexString(ecKey3.getPrivKeyBytes());
-  ECKey ecKey4 = new ECKey(Utils.getRandom());
-  byte[] participateAddress = ecKey4.getAddress();
-  String participateKey = ByteArray.toHexString(ecKey4.getPrivKeyBytes());
   private long multiSignFee = Configuration.getByPath("testng.conf")
       .getLong("defaultParameter.multiSignFee");
   private long updateAccountPermissionFee = Configuration.getByPath("testng.conf")
       .getLong("defaultParameter.updateAccountPermissionFee");
   private ManagedChannel channelFull = null;
   private WalletGrpc.WalletBlockingStub blockingStubFull = null;
+  private ManagedChannel channelFull1 = null;
+  private WalletGrpc.WalletBlockingStub blockingStubFull1 = null;
   private String fullnode = Configuration.getByPath("testng.conf")
       .getStringList("fullnode.ip.list").get(0);
+  private String fullnode1 = Configuration.getByPath("testng.conf")
+      .getStringList("fullnode.ip.list").get(1);
 
   @BeforeSuite
   public void beforeSuite() {
@@ -75,16 +77,20 @@ public class WalletTestMutiSign011 {
    * constructor.
    */
 
-  @BeforeClass(enabled = false)
+  @BeforeClass(enabled = true)
   public void beforeClass() {
     channelFull = ManagedChannelBuilder.forTarget(fullnode)
         .usePlaintext(true)
         .build();
+    channelFull1 = ManagedChannelBuilder.forTarget(fullnode1)
+        .usePlaintext(true)
+        .build();
     blockingStubFull = WalletGrpc.newBlockingStub(channelFull);
+    blockingStubFull1 = WalletGrpc.newBlockingStub(channelFull1);
   }
 
-  @Test(enabled = false)
-  public void testMutiSign1UpdatePermission() {
+  @Test(enabled = true, threadPoolSize = 1, invocationCount = 1)
+  public void testMutiSignForSmartContract() {
     ecKey1 = new ECKey(Utils.getRandom());
     manager1Address = ecKey1.getAddress();
     manager1Key = ByteArray.toHexString(ecKey1.getPrivKeyBytes());
@@ -98,21 +104,29 @@ public class WalletTestMutiSign011 {
     ownerKey = ByteArray.toHexString(ecKey3.getPrivKeyBytes());
     PublicMethed.printAddress(ownerKey);
 
-    long needCoin = updateAccountPermissionFee * 2 + multiSignFee * 3;
+    long needcoin = updateAccountPermissionFee + multiSignFee * 4;
 
     Assert.assertTrue(
-        PublicMethed.sendcoin(ownerAddress, needCoin, fromAddress, testKey002,
+        PublicMethed.sendcoin(ownerAddress, needcoin + 100000000L, fromAddress, testKey002,
             blockingStubFull));
+    Assert.assertTrue(PublicMethed
+        .freezeBalanceForReceiver(fromAddress, 1000000000, 0, 0, ByteString.copyFrom(ownerAddress),
+            testKey002, blockingStubFull));
+    Assert.assertTrue(PublicMethed
+        .freezeBalanceForReceiver(fromAddress, 1000000000, 0, 1, ByteString.copyFrom(ownerAddress),
+            testKey002, blockingStubFull));
+
     PublicMethed.waitProduceNextBlock(blockingStubFull);
-    Long balanceBefore = PublicMethed.queryAccount(ownerAddress, blockingStubFull).getBalance();
+
+    Long balanceBefore = PublicMethed.queryAccount(ownerAddress, blockingStubFull)
+        .getBalance();
     logger.info("balanceBefore: " + balanceBefore);
 
     permissionKeyString[0] = manager1Key;
     permissionKeyString[1] = manager2Key;
+    PublicMethed.waitProduceNextBlock(blockingStubFull);
     ownerKeyString[0] = ownerKey;
     ownerKeyString[1] = manager1Key;
-    String[] permissionKeyString1 = new String[1];
-    permissionKeyString1[0] = ownerKey;
     accountPermissionJson =
         "{\"owner_permission\":{\"type\":0,\"permission_name\":\"owner\",\"threshold\":2,\"keys\":["
             + "{\"address\":\"" + PublicMethed.getAddressString(manager1Key) + "\",\"weight\":1},"
@@ -124,73 +138,62 @@ public class WalletTestMutiSign011 {
             + "{\"address\":\"" + PublicMethed.getAddressString(manager1Key) + "\",\"weight\":1},"
             + "{\"address\":\"" + PublicMethed.getAddressString(manager2Key) + "\",\"weight\":1}"
             + "]}]}";
-
     logger.info(accountPermissionJson);
-    String txid = PublicMethedForMutiSign
-        .accountPermissionUpdateForTransactionId(accountPermissionJson, ownerAddress, ownerKey,
-            blockingStubFull, ownerKeyString);
+    PublicMethedForMutiSign.accountPermissionUpdate(accountPermissionJson, ownerAddress, ownerKey,
+        blockingStubFull, ownerKeyString);
+
+    Random rand = new Random();
+    Integer randNum = rand.nextInt(30) + 1;
+    randNum = rand.nextInt(4000);
+
+    Long maxFeeLimit = 1000000000L;
+    String contractName = "StorageAndCpu" + Integer.toString(randNum);
+    String code = Configuration.getByPath("testng.conf")
+        .getString("code.code_TestStorageAndCpu_storageAndCpu");
+    String abi = Configuration.getByPath("testng.conf")
+        .getString("abi.abi_TestStorageAndCpu_storageAndCpu");
+    byte[] contractAddress = PublicMethedForMutiSign.deployContract1(contractName, abi, code,
+        "", maxFeeLimit,
+        0L, 100, null, ownerKey, ownerAddress, blockingStubFull, 0, ownerKeyString);
+
     PublicMethed.waitProduceNextBlock(blockingStubFull);
-
-    Optional<TransactionInfo> infoById = PublicMethed
-        .getTransactionInfoById(txid, blockingStubFull);
-
+    SmartContract smartContract = PublicMethed.getContract(contractAddress, blockingStubFull);
+    Assert.assertTrue(smartContract.getAbi().toString() != null);
+    String txid;
+    String initParmes = "\"" + "930" + "\"";
+    txid = PublicMethedForMutiSign.triggerContract1(contractAddress,
+        "testUseCpu(uint256)", initParmes, false,
+        0, maxFeeLimit, ownerAddress, ownerKey, blockingStubFull, 0, ownerKeyString);
+    PublicMethed.waitProduceNextBlock(blockingStubFull);
+    PublicMethed.getTransactionById(txid, blockingStubFull);
+    infoById = PublicMethed.getTransactionInfoById(txid, blockingStubFull);
+    Assert.assertTrue(
+        PublicMethedForMutiSign.updateSettingWithPermissionId(contractAddress, 50, ownerKey,
+            ownerAddress, 0, blockingStubFull, ownerKeyString));
+    Assert.assertTrue(
+        PublicMethedForMutiSign.updateEnergyLimitWithPermissionId(contractAddress, 50, ownerKey,
+            ownerAddress, 0, blockingStubFull, ownerKeyString));
+    PublicMethed.waitProduceNextBlock(blockingStubFull);
     long balanceAfter = PublicMethed.queryAccount(ownerAddress, blockingStubFull).getBalance();
-    long energyFee = infoById.get().getReceipt().getEnergyFee();
-    long netFee = infoById.get().getReceipt().getNetFee();
-    long fee = infoById.get().getFee();
-
     logger.info("balanceAfter: " + balanceAfter);
-    logger.info("energyFee: " + energyFee);
-    logger.info("netFee: " + netFee);
-    logger.info("fee: " + fee);
 
-    Assert.assertEquals(balanceBefore - balanceAfter, fee);
-    Assert.assertEquals(fee, energyFee + netFee + updateAccountPermissionFee);
-
-    balanceBefore = balanceAfter;
-    String accountPermissionJson1 =
-        "{\"owner_permission\":{\"type\":0,\"permission_name\":\"owner\",\"threshold\":2,\"keys\":["
-            + "{\"address\":\"" + PublicMethed.getAddressString(manager1Key) + "\",\"weight\":1},"
-            + "{\"address\":\"" + PublicMethed.getAddressString(ownerKey)
-            + "\",\"weight\":1}]},"
-            + "\"active_permissions\":[{\"type\":2,\"permission_name\":\"active0\",\"threshold\":2,"
-            + "\"operations\":\"" + operations + "\","
-            + "\"keys\":["
-            + "{\"address\":\"" + PublicMethed.getAddressString(manager1Key) + "\",\"weight\":1},"
-            + "{\"address\":\"" + PublicMethed.getAddressString(ownerKey) + "\",\"weight\":1}"
-            + "]}]}";
-    txid = PublicMethedForMutiSign
-        .accountPermissionUpdateForTransactionId1(accountPermissionJson1, ownerAddress, ownerKey,
-            blockingStubFull, 2, permissionKeyString);
-    PublicMethed.waitProduceNextBlock(blockingStubFull);
-
-    Assert.assertNotNull(txid);
-
-    infoById = PublicMethed
-        .getTransactionInfoById(txid, blockingStubFull);
-    balanceAfter = PublicMethed.queryAccount(ownerAddress, blockingStubFull).getBalance();
-    energyFee = infoById.get().getReceipt().getEnergyFee();
-    netFee = infoById.get().getReceipt().getNetFee();
-    fee = infoById.get().getFee();
-
-    logger.info("balanceAfter: " + balanceAfter);
-    logger.info("energyFee: " + energyFee);
-    logger.info("netFee: " + netFee);
-    logger.info("fee: " + fee);
-
-
+    Assert.assertTrue(
+        PublicMethed.unFreezeBalance(fromAddress, testKey002, 0, ownerAddress, blockingStubFull));
+    Assert.assertTrue(
+        PublicMethed.unFreezeBalance(fromAddress, testKey002, 1, ownerAddress, blockingStubFull));
   }
-
 
   /**
    * constructor.
    */
-  @AfterClass(enabled = false)
+
+  @AfterClass
   public void shutdown() throws InterruptedException {
     if (channelFull != null) {
       channelFull.shutdown().awaitTermination(5, TimeUnit.SECONDS);
     }
+    if (channelFull1 != null) {
+      channelFull1.shutdown().awaitTermination(5, TimeUnit.SECONDS);
+    }
   }
 }
-
-
