+++ framework/src/test/resources/testng.conf
@@ -144,7 +144,7 @@ defaultParameter = {
   httpConnectionTimeout = 19000
   httpSoTimeout = 18000
   createWitnessAmount = 9999000000
-  operations = 7fff1fc0033e0900000000000000000000000000000000000000000000000000
+  operations = 7fff1fc0033e3300000000000000000000000000000000000000000000000000
   delayTransactionFee = 100000
   cancleDelayTransactionFee = 50000
   solidityCompilerVersion = "v5"
